refactor: 관리자 레이아웃 -> 루트 레이아웃으로 분리 및 적용
