+++ frontend/src/features/admin/index.ts
@@ -1,8 +1,5 @@
 export { default as AdminContent } from './components/AdminContent';
-export { default as AdminHeader } from './components/AdminHeader';
-export { default as AdminLayout } from './components/AdminLayout';
-export { default as AdminNavItem } from './components/AdminNavItem';
-export { default as AdminSidebar } from './components/AdminSidebar';
+export { default as AdminSidebar } from '../../layout/admin/AdminSidebar';
 export { default as AdminTable } from './components/AdminTable';
 export { default as AdminReportTableBody } from './components/AdminReportTableBody';
 export { default as AdminUnverifiedReviewsTableBody } from './components/AdminUnverifiedReviewsTableBody';
