+++ frontend/src/features/auth/models/auth.model.ts
@@ -1,3 +1,9 @@
+export interface IAuthProps {
+  email?: string;
+  nickname?: string;
+  password?: string;
+}
+
 export interface IUser {
   email: string;
   nickname: string;
