+++ frontend/src/pages/Join/JoinStep1Email.tsx
@@ -1,7 +1,7 @@
 import { useForm } from 'react-hook-form';
 import styled from 'styled-components';
 
-import Layout from '@/layout/Layout';
+import Layout from '@/layout/user/Layout';
 import {
   IUserRegistration,
   JoinTemplate,
