+++ frontend/src/pages/Mypage/MyReviews.tsx
@@ -1,4 +1,4 @@
-import Layout from '@/layout/Layout';
+import Layout from '@/layout/user/Layout';
 import { ReviewList, formatReviews, useMyReviews } from '@/features/reviews';
 
 const MyReviews = () => {
