+++ frontend/src/pages/Review/ReviewList.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import Layout from '@/layout/Layout';
+import Layout from '@/layout/user/Layout';
 import {
   ReviewList,
   useInfiniteReviewListWithParams,
